Extract JavascriptType and JavaToJavascriptTypeConversionUtil to separate module
