+++ nrich-javascript/README.MD
@@ -1,0 +1,12 @@
+# nrich-javascript
+
+[![Maven Central](https://maven-badges.herokuapp.com/maven-central/net.croz.nrich/nrich-javascript/badge.svg?color=blue)](https://maven-badges.herokuapp.com/maven-central/net.croz.nrich/nrich-javascript)
+
+## Overview
+
+nrich-javascript is a module that contains classes for conversion between Java class and its Javascript representation. It is used internally by `nrich-form-configuration` and `nrich-registry`
+modules.
+
+## Usage
+
+This module is primary internal to nrich. But if users want to use it for converting Java type to Javascript type the class `JavaToJavascriptTypeConversionUtil` should be used.
