+++ nrich-javascript/build.gradle
@@ -1,0 +1,11 @@
+description = "Nrich Javascript"
+
+dependencies {
+  implementation "com.fasterxml.jackson.core:jackson-annotations"
+
+  testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine"
+
+  testImplementation "org.assertj:assertj-core"
+  testImplementation "org.junit.jupiter:junit-jupiter-api"
+  testImplementation "org.junit.jupiter:junit-jupiter-params"
+}
